% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addScaleBar.R
\name{addScaleBar}
\alias{addScaleBar}
\title{Add a horizontal scale bar to image}
\usage{
addScaleBar(
  image = NULL,
  length_per_pixel_in_um = NULL,
  distance_from_border = 20
)
}
\arguments{
\item{image}{A 2 or 3 dimensional array which could also be of the
formal EBImage image class.}

\item{length_per_pixel_in_um}{A number representing thelength per pixel in um.}

\item{distance_from_border}{A number being the distance in pixels of the
scale bar from the right image border.}
}
\value{
An array (image with added scale bar).
}
\description{
`addScaleBar()` adds a scale bar to an image whose size depends on
the metadata of the image
}
