% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCziMetadata_AxioImager.R
\name{readCziMetadata_AxioImager}
\alias{readCziMetadata_AxioImager}
\title{Check for specific values in metadata for all other microscope types}
\usage{
readCziMetadata_AxioImager(
  metadata = metadata,
  number_of_channels = number_of_channels
)
}
\arguments{
\item{metadata}{A character (loaded metadata in a string)}

\item{number_of_channels}{A number}
}
\description{
Check for specific values in metadata for all other microscope types
}
\keyword{internal}
