% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCziMetadata_LSM.R
\name{readCziMetadata_LSM}
\alias{readCziMetadata_LSM}
\title{Check for apotome-specific values in metadata}
\usage{
readCziMetadata_LSM(input_file = NULL, number_of_channels = number_of_channels)
}
\arguments{
\item{number_of_channels}{A number}

\item{metadata}{A character (loaded metadata in a string)}
}
\description{
Check for apotome-specific values in metadata
}
\keyword{internal}
