% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCziMetadata_LSM_PMT.R
\name{readCziMetadata_LSM_PMT}
\alias{readCziMetadata_LSM_PMT}
\title{Check for apotome-specific values in metadata}
\usage{
readCziMetadata_LSM_PMT(
  metadata = metadata,
  number_of_channels = number_of_channels
)
}
\arguments{
\item{metadata}{A character (loaded metadata in a string)}

\item{number_of_channels}{A number}
}
\description{
Check for apotome-specific values in metadata
}
\keyword{internal}
