% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resizeImage.R
\name{resizeImage}
\alias{resizeImage}
\title{Resize an image}
\usage{
resizeImage(image = NULL, resize_factor = 1)
}
\arguments{
\item{image}{A one- to three-dimensional array of numbers between 0 and 1.}

\item{resize_factor}{A number being the factor for resizing an image.}
}
\value{
An array (resized image).
}
\description{
`resizeImage()` resizes an image with a given scaling factor
}
