% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertCziToTif.R
\name{convertCziToTif}
\alias{convertCziToTif}
\alias{convertczitotif}
\alias{convertCzitotif}
\alias{convertcziTotif}
\alias{convertczitoTif}
\alias{convertCziTotif}
\alias{convertcziToTif}
\title{convertCziToTif}
\usage{
convertCziToTif(
  input_file = NULL,
  convert_all_slices = TRUE,
  stack_image = TRUE,
  stack_method = "average",
  higher_contrast_slices = TRUE,
  higher_contrast_stack = TRUE
)
}
\arguments{
\item{input_file}{A character (path to czi file to be converted)}

\item{convert_all_slices}{A logical (T/F for converting and saving all z-slices)}

\item{stack_image}{A logical (T/F for calculating a z-stack)}

\item{stack_method}{A character (method for stacking the images)}

\item{higher_contrast_slices}{A logical (T/F for enhancing the contrast of the slices)}

\item{higher_contrast_stack}{A logical (T/F for enhancing the constrast of the z-stack)}
}
\description{
Imports czi file and save them as tif
}
\details{
This functions saves a multidimensional array (image) as tifs
and stacks/edits the images if required.
}
\author{
Kai Budde
}
