% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCzi.R
\name{readCzi}
\alias{readCzi}
\alias{readczi}
\alias{readCZI}
\title{readCzi}
\usage{
readCzi(input_file = NULL)
}
\arguments{
\item{input_file}{A character (path to czi file to be converted)}
}
\description{
Imports czi files and save the metadata
}
\details{
The function returns a multidimensional array with pixel
identities from 0 to 1. The order of the array's dimensions are:
row (y), column (x), channel (red, green, blue), slices (z)
}
\author{
Kai Budde
}
