% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCziMetadata_Apotome.R
\name{readCziMetadata_Apotome}
\alias{readCziMetadata_Apotome}
\title{Check for apotome-specific values in metadata}
\usage{
readCziMetadata_Apotome(
  metadata = NULL,
  number_of_channels = NULL,
  number_of_tracks = NULL
)
}
\arguments{
\item{metadata}{A character (loaded metadata in a string)}

\item{number_of_channels}{A number}
}
\description{
Check for apotome-specific values in metadata
}
\keyword{internal}
