% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCziMetadata_AxioImager.R
\name{readCziMetadata_AxioImager}
\alias{readCziMetadata_AxioImager}
\title{Check for specific Axio-imager values in metadata}
\usage{
readCziMetadata_AxioImager(
  metadata = NULL,
  number_of_channels = NULL,
  number_of_tracks = NULL
)
}
\arguments{
\item{metadata}{A character (loaded metadata in a string)}

\item{number_of_channels}{A number}
}
\description{
Check for specific Axio-imager values in metadata
}
\keyword{internal}
