% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCziMetadata_AxioImager.R
\name{readCziMetadata_AxioImager}
\alias{readCziMetadata_AxioImager}
\title{Check for specific Axio-imager values in metadata}
\usage{
readCziMetadata_AxioImager(
  metadata = NULL,
  number_of_channels = NULL,
  number_of_tracks = NULL
)
}
\arguments{
\item{metadata}{Loaded metadata as a string.}

\item{number_of_channels}{A number representing the number of channels
(color layers).}
}
\description{
Check for specific Axio-imager values in metadata
}
\keyword{internal}
