% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCziMetadata_LSM.R
\name{readCziMetadata_LSM}
\alias{readCziMetadata_LSM}
\title{Check for LSM-specific values in metadata}
\usage{
readCziMetadata_LSM(
  metadata = NULL,
  number_of_channels = NULL,
  number_of_tracks = NULL
)
}
\arguments{
\item{metadata}{A character (loaded metadata in a string)}

\item{number_of_channels}{A number}

\item{number_of_tracks}{A number}
}
\description{
Check for LSM-specific values in metadata
}
\keyword{internal}
