% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resizeImage.R
\name{resizeImage}
\alias{resizeImage}
\alias{resizeimage}
\title{resizeImage}
\usage{
resizeImage(image = NULL, resize_factor = 1)
}
\arguments{
\item{image}{An one to three-dimensional array of numbers between 0 and 1}

\item{resize_factor}{A number (factor for resizing an image)}
}
\description{
Resizes an image
}
\details{
Resizes an image with a given scaling factor
}
\author{
Kai Budde
}
