% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackLayers.R
\name{stackLayers}
\alias{stackLayers}
\alias{stacklayers}
\alias{StackLayers}
\title{stackLayers}
\usage{
stackLayers(image_data = NULL, stack_method = "average")
}
\arguments{
\item{image_data}{An array of EBImage object}

\item{stack_method}{A character (method for stacking the images: average,
maxprojection)}
}
\description{
Stacks z-layers
}
\details{
This functions saves a multidimensional array (image) as tifs
and stacks/edits the images if required.
}
\author{
Kai Budde
}
